refactor: LocalMemoryBoardStatus의 검증 책임을 Board로 이동
